LayoutTests/http/tests/appcache/local-content-expected.txt

CONSOLE MESSAGE: line 0: Not allowed to load local resource: stdio.h
Test that documents loaded from application cache don't get access to local resources accidentally.

Should say SUCCESS:

SUCCESS
